Create events controller
